Rename 'foo-procedures' to 'banana-procedures'
These procedures use the Bananation mission model, not the Foo mission model
