+++ procedural/examples/banana-procedures/README.md
@@ -1,0 +1,13 @@
+# banana-procedures
+
+This is a project for test examples. It does provide valid examples
+of goals and constraints, but they aren't particularly helpful for users.
+
+## Building
+
+At repo top level:
+
+```bash
+./gradlew build
+./gradlew :procedural:examples:banana-procedures:buildAllSchedulingProcedureJars
+```
