+++ procedural/examples/banana-procedures/src/main/java/gov/nasa/ammos/aerie/procedural/examples/bananaprocedures/Helper.java
@@ -1,0 +1,7 @@
+package gov.nasa.ammos.aerie.procedural.examples.bananaprocedures;
+
+public class Helper {
+  public static String greeting() {
+    return "hello from util";
+  }
+}
