+++ procedural/examples/banana-procedures/src/main/java/gov/nasa/ammos/aerie/procedural/examples/bananaprocedures/constraints/Hello.java
@@ -1,0 +1,9 @@
+package gov.nasa.ammos.aerie.procedural.examples.bananaprocedures.constraints;
+
+import gov.nasa.ammos.aerie.procedural.examples.bananaprocedures.Helper;
+
+class Hello {
+  public static void main(String[] args) {
+    System.out.println(Helper.greeting());
+  }
+}
