+++ procedural/examples/banana-procedures/src/main/java/gov/nasa/ammos/aerie/procedural/examples/bananaprocedures/package-info.java
@@ -1,0 +1,5 @@
+@WithMappers(BasicValueMappers.class)
+package gov.nasa.ammos.aerie.procedural.examples.bananaprocedures;
+
+import gov.nasa.jpl.aerie.contrib.serialization.rulesets.BasicValueMappers;
+import gov.nasa.ammos.aerie.procedural.scheduling.annotations.WithMappers;
