--- procedural/examples/foo-procedures/README.md
@@ -1,13 +1,0 @@
-# foo-procedures
-
-This is a project for test examples. It does provide valid examples
-of goals and constraints, but they aren't particularly helpful for users.
-
-## Building
-
-At repo top level:
-
-```bash
-./gradlew build
-./gradlew :procedural:examples:foo-procedures:buildAllSchedulingProcedureJars
-```
