--- procedural/examples/foo-procedures/src/main/java/gov/nasa/ammos/aerie/procedural/examples/fooprocedures/Helper.java
@@ -1,7 +1,0 @@
-package gov.nasa.ammos.aerie.procedural.examples.fooprocedures;
-
-public class Helper {
-  public static String greeting() {
-    return "hello from util";
-  }
-}
