--- procedural/examples/foo-procedures/src/main/java/gov/nasa/ammos/aerie/procedural/examples/fooprocedures/constraints/Hello.java
@@ -1,9 +1,0 @@
-package gov.nasa.ammos.aerie.procedural.examples.fooprocedures.constraints;
-
-import gov.nasa.ammos.aerie.procedural.examples.fooprocedures.Helper;
-
-class Hello {
-  public static void main(String[] args) {
-    System.out.println(Helper.greeting());
-  }
-}
