--- procedural/examples/foo-procedures/src/main/java/gov/nasa/ammos/aerie/procedural/examples/fooprocedures/package-info.java
@@ -1,5 +1,0 @@
-@WithMappers(BasicValueMappers.class)
-package gov.nasa.ammos.aerie.procedural.examples.fooprocedures;
-
-import gov.nasa.jpl.aerie.contrib.serialization.rulesets.BasicValueMappers;
-import gov.nasa.ammos.aerie.procedural.scheduling.annotations.WithMappers;
